Implemented Register class

+++ Code/infra/types.h
@@ -3,20 +3,20 @@
 
 #include <cstdint>
 
-using int8 = int8_t;
-using int16 = int16_t;
-using int32 = int32_t;
-using int64 = int64_t;
-using uint8 = uint8_t;
-using uint16 = uint16_t;
-using uint32 = uint32_t;
-using uint64 = uint64_t;
+using int8    = int8_t;
+using int16   = int16_t;
+using int32   = int32_t;
+using int64   = int64_t;
+using uint8   = uint8_t;
+using uint16  = uint16_t;
+using uint32  = uint32_t;
+using uint64  = uint64_t;
 using float32 = float;
 using float64 = double;
 
 // semantics
-using Addr = uint32;
-using Cycles = uint64;
+using Addr   = uint32;
+using Cycles = uint32;
 
 static const uint8  NO_VAL8  = 0xba;
 static const uint16 NO_VAL16 = 0xf00d;
